Revert "fix: uint -> int pk"
This reverts commit 18b2ebaecbbbfff3aa44ffc7d05505135ecbb27a.

+++ internal/app/handler/material.go
@@ -47,7 +47,7 @@ func (h *Handler) AddMaterialToExperiment(ctx *gin.Context) {
 		return
 	}
 
-	err = h.Repository.AddMaterialToExperiment(experimentId, materialId)
+	err = h.Repository.AddMaterialToExperiment(experimentId, uint(materialId))
 	if err != nil {
 		h.errorHandler(ctx, http.StatusInternalServerError, err)
 		return
